requirements.txt
streamlit
tensorflow
numpy
pillow
opencv-python-headless
scikit-learn
matplotlib
joblib